fastapi-backend/requirements.txt
acres==0.3.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorclass==2.2.2
configobj==5.0.9
configparser==7.2.0
cryptography==44.0.1
decorator==5.2.1
easygui==0.98.3
et_xmlfile==2.0.0
etelemetry==0.3.1
fastapi==0.115.11
ffmpeg-python==0.2.0
filelock==3.18.0
frontend==0.0.3
future==1.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
isodate==0.6.1
itsdangerous==2.2.0
joblib==1.4.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
looseversion==1.3.0
lxml==5.3.2
msgpack==1.1.0
msoffcrypto-tool==5.4.2
mutagen==1.47.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numba==0.61.0
numpy==2.1.3
olefile==0.47
oletools==0.60.2
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pcodedmp==1.2.6
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
prov==2.0.1
ptyprocess==0.7.0
puremagic==1.28
pyClamd==0.4.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
PyExifTool==0.5.6
pymediainfo==7.0.1
PyMuPDF==1.25.5
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-multipart==0.0.20
pytils==0.4.3
pytz==2025.1
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.46.1
stegpy==0.0.1
threadpoolctl==3.6.0
tifffile==2025.2.18
tools==0.1.9
traits==7.0.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
yara-python==4.5.1
